// apps/service/src/common/interface.ts
export class IElectricMeterImageResponse {
  electricMeterReading: string;
  manufacturer: string;
  model: string;
  serialNumber: string;
  voltage: string;
  current: string;
  frequency: string;
  powerFactor: string;
  temperature: string;
  phase: string;
  installationYear: string;
  locationMarking: string;
}